LangChain-Guide/pyproject.toml
[tool.poetry]
name = "langchain-guide"
version = "0.1.0"
description = ""
authors = ["abdulrehman1004 <rehman.abdul008844@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "^1.0.1"
langchain = "^0.3.6"
langchain-community = "^0.3.4"
langchain-google-genai = "^2.0.4"
langchain-huggingface = "^0.1.2"
langchain-qdrant = "^0.1.4"
transformers = "^4.46.1"
qdrant-client = "^1.12.1"
pypdf = "^5.1.0"
docx2txt = "^0.8"
pandas = "^2.2.3"
langchain-pinecone = "^0.2.0"
pinecone-notebooks = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
